requirements.txt
mysql-connector-python 
faker
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
itsdangerous